フロント側をゴリゴリにリファクタした
