+++ app/assets/stylesheets/application.sass.scss
@@ -5,7 +5,6 @@
 // utility
 @use './utility/colors';
 @use './utility/responsive';
-@use './utility/common';
 
 // React各コンポーネント用scss
 @use './pages/pages.scss';
