+++ app/assets/stylesheets/templates/Container/_style.scss
@@ -1,4 +1,4 @@
-@use 'colors' as *;
+@use '../../utility/colors' as *;
 
 .l-container {
   background-color: $white;
