--- app/assets/stylesheets/templates/LoginForm/_style.scss
@@ -1,27 +1,0 @@
-@use '../../utility/colors' as *;
-
-.t-login{
-  margin: 30px auto;
-  border: 2px solid $lightBrown;
-  padding: 24px;
-
-  &__input {
-    margin-bottom: 20px;
-    label {
-      display: block;
-      margin-bottom: 5px;
-    }
-
-    input {
-      width: 100%;
-      padding: 8px;
-      font-size: 1rem;
-      line-height: 1.5;
-      border: 1px solid $lightBrown;
-      border-radius: 4px;
-    }
-  }
-  &__button{
-    text-align: center;
-  }
-}
