+++ app/assets/stylesheets/templates/Page/_style.scss
@@ -1,0 +1,5 @@
+@use '../../utility/colors' as *;
+
+.t-page {
+  margin: 50px 70px;
+}
