+++ app/assets/stylesheets/templates/_templates.scss
@@ -1,4 +1,6 @@
 // style定義禁止
 // フォルダ生やしたら適宜パスの追加
 @use '../templates/Header/style.scss' as Header;
-@use '../templates/LoginForm/style.scss' as LoginForm;
+@use '../templates/Form/style.scss' as Form;
+@use '../templates/Page/style.scss' as Page;
+@use '../templates/Container/style.scss' as Container;
