+++ app/javascript/application.js
@@ -1,19 +1,18 @@
-// NOTIE : このファイルは触らないでください
 // Entry point for the build script in your package.json
 import React from "react";
 import { createRoot } from "react-dom/client";
 import Top from "./pages/Top";
+import { AuthLogin } from "./pages/AuthLogin";
 import { AuthRegister } from "./pages/AuthRegister";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-/* import { Header } from "./templates/Header"; */
 
 const root = createRoot(document.getElementById("root"));
 root.render(
   <BrowserRouter>
-    {/* <Header /> */} {/* application.jsにHeaderを入れるとレイアウトが崩れるので、一旦コメントアウト */}
     <Routes>
       <Route path="/" element={<Top />} />
       <Route exact path="/register" element={<AuthRegister />} />
+      <Route exact path="/login" element={<AuthLogin />} />
     </Routes>
   </BrowserRouter>
 );
