+++ app/javascript/pages/AuthRegister/index.tsx
@@ -4,6 +4,10 @@ import { useNavigate } from "react-router-dom";
 
 import { Header } from "../../templates/Header";
 import prefectures from "../../utility/prefecture";
+import { H2Header } from "../../parts/H2Header";
+import { FormTemplate } from "../../templates/FormTemplate";
+import { Button } from "../../parts/Button";
+import { PageTemplate } from "../../templates/PageTemplate";
 
 
 type Inputs = {
@@ -73,17 +77,17 @@ export const AuthRegister: React.FC = () => {
   return (
     <div className="l-container">
       <Header />
-      <div className="p-signup">
-        <h1 className="p-signup__title">会員登録</h1>
+      <PageTemplate>
+        <H2Header>会員登録</H2Header>
         <p className="p-signup__description">下記の情報を入力して、「新規会員登録する」ボタンを押してください。</p>
-        <form className="p-signup__form" onSubmit={handleSubmit(onSubmit)}>
-          <div className="p-signup__form--label-with-validation-message">
+        <FormTemplate onSubmit={handleSubmit(onSubmit)}>
+          <div className="t-form__input p-signup__form--label-with-validation-message">
             <label htmlFor="username">氏名</label>
             {errors.username?.message && <span>{errors.username.message}</span>}
           </div>
           <input type="text" name="username" ref={register({ required: "ユーザー名は必須です。" })} />
 
-          <div className="p-signup__form--label-with-validation-message">
+          <div className="t-form__input p-signup__form--label-with-validation-message">
             <label htmlFor="email">メールアドレス</label>
             {errors.email?.message && <span>{errors.email.message}</span>}
           </div>
@@ -94,7 +98,7 @@ export const AuthRegister: React.FC = () => {
             }
           })} />
 
-          <div className="p-signup__form--label-with-validation-message">
+          <div className="t-form__input p-signup__form--label-with-validation-message">
             <label htmlFor="password">パスワード</label>
             {errors.password?.message && <span>{errors.password.message}</span>}
           </div>
@@ -103,8 +107,8 @@ export const AuthRegister: React.FC = () => {
             minLength: { value: 6, message: "パスワードは6文字以上入力してください。" }
           })} />
 
-          <div className="p-signup__form--label-with-validation-message">
-            <label htmlFor="passwordConfirmation">確認のため再度入力してください</label>
+          <div className="t-form__input p-signup__form--label-with-validation-message">
+            <label htmlFor="passwordConfirmation">パスワード(確認用)</label>
             {errors.passwordConfirmation?.message && <span>{errors.passwordConfirmation.message}</span>}
           </div>
           <input type="password" name="passwordConfirmation" ref={register({
@@ -112,16 +116,16 @@ export const AuthRegister: React.FC = () => {
             validate: (value: string) => value === watch("password") || "パスワードが一致しません。"
           })} />
 
-          <div className="p-signup__form--label-with-validation-message">
+          <div className="t-form__input p-signup__form--label-with-validation-message">
             <label htmlFor="userType">ユーザータイプ</label>
             {errors.userType?.message && <span>{errors.userType.message}</span>}
           </div>
-          <div className="p-signup__form--user-type">
+          <div className="t-form__input p-signup__form--user-type">
             <input type="radio" id="offerer" value={1} name="userType" ref={register({ required: "ユーザータイプを選択してください。" })} /><label htmlFor="offerer">募集者</label>
             <input type="radio" id="seeker" value={2} name="userType" ref={register({ required: "ユーザータイプを選択してください。" })} /><label htmlFor="seeker">応募者</label>
           </div>
 
-          <div className="p-signup__form--label-with-validation-message">
+          <div className="t-form__input p-signup__form--label-with-validation-message">
             <label htmlFor="zipcode">郵便番号</label>
             {errors.zipcode?.message && <span>{errors.zipcode.message}</span>}
           </div>
@@ -132,11 +136,7 @@ export const AuthRegister: React.FC = () => {
               required: "郵便番号は必須です。",
               pattern: { value: /^\d{7}$/, message: "有効な郵便番号7桁を半角入力してください (例: 1234567) " }
             })} />
-
-          {/* <div className="p-signup__form--address-block"> */}
-
-          {/* <div className="p-signup__form--prefecture"> */}
-          <div className="p-signup__form--label-with-validation-message">
+          <div className="t-form__input p-signup__form--label-with-validation-message">
             {/* <div className="p-signup__form--prefecture"> */}
             <label htmlFor="prefecture">都道府県</label>
             {errors.prefecture?.message && <span>{errors.prefecture.message}</span>}
@@ -147,23 +147,17 @@ export const AuthRegister: React.FC = () => {
               <option value={prefecture} key={prefecture}>{prefecture}</option>
             ))}
           </select>
-          {/* </div> */}
-
-          {/* <div className="p-signup__form--city"> */}
-          <div className="p-signup__form--label-with-validation-message">
+          <div className="t-form__input p-signup__form--label-with-validation-message">
             <label htmlFor="city">市区町村</label>
             {errors.city?.message && <span>{errors.city.message}</span>}
           </div>
           <input type="text" name="city" ref={register({ required: "市区町村を入力してください。" })} />
-          {/* </div> */}
-
-          <div className="p-signup__form--label-with-validation-message">
+          <div className="t-form__input p-signup__form--label-with-validation-message">
             <label htmlFor="block">番地・建物名など</label>
             {errors.block?.message && <span>{errors.block.message}</span>}
           </div>
           <input type="text" name="block" ref={register({ required: "番地や建物名を入力してください。" })} />
-
-          <div className="p-signup__form--label-with-validation-message">
+          <div className="t-form__input p-signup__form--label-with-validation-message">
             <label htmlFor="phoneNumber">電話番号</label>
             {errors.phoneNumber?.message && <span>{errors.phoneNumber.message}</span>}
           </div>
@@ -172,9 +166,11 @@ export const AuthRegister: React.FC = () => {
             pattern: { value: /^\d+$/, message: "電話番号は半角数字のみで入力してください。" }
           })} />
 
-          <button type="submit" disabled={!isValid || isSubmitting}>新規会員登録する</button>
-        </form >
-      </div >
+          <div className="t-login__button">
+            <Button type="submit" disabled={!isValid || isSubmitting}>新規会員登録する</Button>
+          </div>
+        </FormTemplate>
+      </PageTemplate>
     </div >
   )
 };
