+++ app/javascript/templates/ContainerTemplate/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+interface ContainerTemplateProps {
+  children: React.ReactNode;
+}
+
+export const ContainerTemplate: React.FC<ContainerTemplateProps> = ({ children }) => {
+  return (
+    <div className="l-container">
+      {children}
+    </div>
+  );
+};
